+++ changelog.md
@@ -1,6 +1,6 @@
 ## Changes in version 0.7.1.0
 
-  * Add `PrimMonad` and `PrimBase` instances for `Lazy.ST` (GHC > 8.2).
+  * Add `PrimMonad` and `PrimBase` instances for `Lazy.ST` (GHC >= 8.2).
   * Add `Eq` instances for `MutableByteArray` and `MutablePrimArray`.
 
   * Allow building with GHC 8.12.